fake-disk: add a new service to create a fake disk
When the environment variable USE_FAKE_DISK is set, and the
FAKE_DISK_FILE does not exist yet, it will be created. FAKE_DISK_FILE
is set to /srv/fake-disk.img by default. When it is expected for the
disk to be persistent over container restarts, it is needed to have the
directory where the disk file is located to be bind-mounted into the
container (a volume providing /srv).

By default the size of the fake disk is set to 10GB. This is a sparse
file allocated with `fallocate`. If a different size is required, the
FAKE_DISK_SIZE environment variable can be used to change this.

The disk file is setup with `losetup` and will be available for use
through /dev/loop* device nodes. Depending on the environment, the
number of the loop device may change. In order to privide a stable name,
FAKE_DISK_DEV can be used. /dev/fake is the default for this.

Signed-off-by: Niels de Vos <ndevos@redhat.com>