gluster-setup: retry in case mounting fails
Currently gluster-setup.sh aborts when 'mount' returns an error. In
certain cases LVM/LVs are not available yet, and mounting the devices is
not possible. The script already tries to account for that, but a
mistake prevents the retry from being reached.

With this change the gluster-setup.sh script exits only when mounting
succeeds, and will do the retry (including 'vgscan --mknods') when a
failure occured.

Signed-off-by: Niels de Vos <ndevos@redhat.com>