gluster-setup: don't start log files with an empty line
While reviewing the logic in the gluster-setup.sh script, it seemed
wrong to start counting lines of 'failed_bricks' at 1. In fact, the
script is correct, it does intentionally skip the 1st line of the file.
The 1st line is empty due to 'echo "" > ...file'. It is more common
practise to create an empty file with '> ...file', but 'shellcheck'
suggests to use 'true > ...file' instead.

After dropping the empty line from the 'failed_bricks' file, it is now
possible to check if the file is empty ('wc -l' will output 0).

Signed-off-by: Niels de Vos <ndevos@redhat.com>