centos: if a "host dev" exists bind mount it over /dev
This is a work around for containerization systems that
have a problem with us using the host's /dev directly but we
need a dynamic dev populated for devices, lvs and gluster-block
support.
Eventually it would be nice to move away from needing the
host's /dev but that day is not here yet.

Signed-off-by: John Mulligan <jmulligan@redhat.com>