gluster-setup: correct check for number of lines in failed_bricks
In case some of the bricks fail to mount, a retry is done. This includes
a check for counting the numbers in the 'failed_bricks' file. The
command that is used will return something like this:

    $ wc -l $GLUSTERFS_LOG_CONT_DIR/failed_bricks
    12 /var/log/glusterfs/container/failed_bricks

Comparing the output with '-gt 0' will never match, as the output is not
convertable to an integer.

The fix was already available in the Fedora/gluster-setup.sh script. It
just needed a little tuning to pass shellcheck.

Signed-off-by: Niels de Vos <ndevos@redhat.com>