gluster-setup: increase cgroup/pids.max to allow many brick processes/threads
docker-1.13.1-94 limits the number of pids (processes+threads) to 4096.
This is too little for an environment where many volumes are served. The
multiplexed brick process has loads of threads, and several of these
processes can be running.

By introducing the CGROUP_PIDS_MAX environment variable, the cgroup for
the pid-limit can be configured for the Gluster container. By default
the value will be set to 'max', as it was before the change in docker.

In case the configuration by the container runtime should not be
modified, set CGROUP_PIDS_MAX to 0.

Bug: https://bugzilla.redhat.com/1698096
See-also: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/Documentation/cgroup-v1/pids.txt
Signed-off-by: Niels de Vos <ndevos@redhat.com>