centos: disable unintentional starting of dmeventd
Masking the dm-event.socket is not sufficient as there will remain a
symlink that triggers running the service under
/etc/systemd/system/sockets.target.wants. For this symlink to be
removed, it is required to disable the dm-event.socket.

In addition to the socket activation, lvm2-monitor.service triggers
starting of dmeventd as well. There is no intention to monitor LVM
within the container, so this service can be disabled and masked too.

Signed-off-by: Niels de Vos <ndevos@redhat.com>