Execute LVM commands on the host instead of in the container
When Gluster runs in a container, it is recommended to run the LVM
commands on the host, and not inside the container. With the new
'exec-on-host' shell script, the commands are executed through 'nsenter'
which expects the /-filesystem of the host to be bind-mounted (as a
volume) on /rootfs. It is also required that the container runs in the
PID-namespace of the host, so that PID 1 refers to the init process on
the host.

If the /-filesystem is not available under /rootfs in the container, or
running 'true' through the 'exec-on-host' script fails, a fallback is
done to execute the command inside the container anyway.

See-also: https://github.com/heketi/heketi/pull/1650
Signed-off-by: Niels de Vos <ndevos@redhat.com>